Adds second layer to the macropad
This has the same keycodes used in the previous KMK firmware, which
means the second layer is only useful to change the rotary encoder into
a mouse wheel instead of a volume knob. This layer should be expanded
for more navigation helpers, and this is why I named it as `NAVIGATION`.