Uses rotary encoder instead of module and includes switch in matrix
The way QMK handles rotary encoder push buttons is to include it in the
switch matrix, which means I needed to add another column. Since this
opens space for an extra switch, I added it here, but it doesn't need to
be a mechanical switch as the others. We can keep the same 3x2 layout of
the main keys.

Using the encoder instead of a module is the right way to go. The
footprint for ALPS EC11E should be used when laying the PCB out because
this seems to be by far the most common format and also matches the ones
I have in stock.