Replaces Pi pico with a RP2040
This is much better for our purposes. It gives us more control, uses
much less space on the PCB and allows us to use a USB-C connector.

All the components around the RP2040 are the ones recommended in the
RP2040 hardware design guide. They not only are recommended, but also
readily available and easy enough to use.