Adds pull-up resistor footprint (component marked as DNP) to the LED data line
QMK has a open-drain WS2812 driver[1], which we can potentially use to
drive the LEDs at 5V. The issue is that, in theory, the RP2040 is not 5V
tolerant, and the QMK docs are very specific about this needing to be
the case. That's why the component is marked as "Do Not Populate", and
we'll leave the footprint there for experimentation.

[1]: https://docs.qmk.fm/drivers/ws2812#open-drain-circuit